STEP2리뷰 &STEP3 alert 반영
