+++ src/js/index.js
@@ -1,5 +1,6 @@
+import { SELECTOR } from './constant.js';
 import { handleFormNameSubmit, handleFormTrialTimesForm } from './handler.js';
-import { $, SELECTOR } from './utils/selector.js';
+import { $ } from './utils/selector.js';
 
 const setEvent = () => {
   $(SELECTOR.CAR_NAME_FORM).addEventListener('submit', (event) =>
