+++ src/js/race.js
@@ -1,6 +1,6 @@
-import { CAR_RACE } from './constant.js';
+import { CAR_RACE, WINNER_MESSAGE } from './constant.js';
 import { renderForward } from './ui/dom.js';
-import { getNumberList, shuffle } from './utils/util.js';
+import { getRandom } from './utils/util.js';
 
 export const startRace = (cars, times) => {
   return new Promise((resolve) => {
@@ -9,13 +9,9 @@ export const startRace = (cars, times) => {
     const intervalId = setInterval(() => {
       count += 1;
       cars.map((car) => {
-        const forwardedStep = shuffle(
-          getNumberList(CAR_RACE.MAX),
-          CAR_RACE.COUNT
-        );
-        car.forwardStep(forwardedStep);
-
+        const forwardedStep = getRandom(CAR_RACE.MIN, CAR_RACE.MAX);
         if (forwardedStep >= CAR_RACE.FORWARD) {
+          car.forwardStep(forwardedStep);
           renderForward(car.name);
         }
       });
@@ -30,6 +26,7 @@ export const startRace = (cars, times) => {
 
 export const getWinner = (raceResult) => {
   const maxStep = Math.max(...raceResult.map((car) => car.step));
+  setTimeout(() => alert(WINNER_MESSAGE), 2000);
   return raceResult
     .filter((car) => car.step === maxStep)
     .map((car) => car.name);
