+++ src/js/utils/selector.js
@@ -1,21 +1,2 @@
 export const $ = (selector) => document.querySelector(selector);
 export const $$ = (selector) => document.querySelectorAll(selector);
-export const SELECTOR = {
-  CAR_NAME_FORM: '.name-form',
-  CAR_NAME_INPUT: '.car-name',
-  CAR_NAME_BUTTON: '.car-name-submit-btn',
-
-  TRIAL_NUMBER_FORM: '.times-form',
-  TRIAL_NUMBER_INPUT: '.trial-number',
-  TRIAL_NUMBER_BUTTON: '.trial-number-submit-btn',
-
-  CAR_PLAYER: '.car-player',
-  FORWARD_ICON: '.forward-icon',
-  SPINNER_ICON: '.spinner',
-
-  RACE_TIMES_SECTION: '.race-times',
-  RACE_PROCESS_COMPONENT: '.race-process-component',
-  RACE_RESULT_COMPONENT: '.race-result-component',
-
-  SPINNER_WRAPPER: '.spinner-wrapper',
-};
