+++ src/js/utils/util.js
@@ -1,11 +1,5 @@
-export const getNumberList = (max) => {
-  return Array(max + 1)
-    .fill()
-    .map((number, idx) => idx);
-};
-
-export const shuffle = (array, count) => {
-  return array.sort(() => Math.random() - 0.5).slice(0, count);
+export const getRandom = (min, max) => {
+  return Math.random() * (max - min) + min;
 };
 
 export const tab =
